+++ sh.d/test-small.sh.d/22-apb-pe.sh
@@ -20,21 +20,21 @@ samtools fastq \
 deps/ART_profiler_illumina/art_profiler_illumina \
     "${OUT_DIR}"/out_pe_art_perl_ "${OUT_DIR}"/ fq
 
-ABP_EXEC \
+APB_EXEC \
     --i-file "${OUT_DIR}"/out_pe.1.fq \
     --read_len "${RLEN}" \
     --o-file1 "${OUT_DIR}"/out_pe_art_cxx_fq_R1.txt \
     --parallel "${PARALLEL}" \
     --i-num_threads 4 \
     --old_behavior
-ABP_EXEC \
+APB_EXEC \
     --i-file "${OUT_DIR}"/out_pe.2.fq \
     --read_len "${RLEN}" \
     --o-file1 "${OUT_DIR}"/out_pe_art_cxx_fq_R2.txt \
     --parallel "${PARALLEL}" \
     --i-num_threads 4 \
     --old_behavior
-ABP_EXEC \
+APB_EXEC \
     --i-file "${OUT_DIR}"/out_pe.sam \
     --read_len "${RLEN}" \
     --is_pe \
@@ -50,6 +50,39 @@ cmp "${OUT_DIR}"/out_pe_art_cxx_fq_R2.txt "${OUT_DIR}"/out_pe_art_perl_R2.txt
 cmp "${OUT_DIR}"/out_pe_art_cxx_sam_R1.txt "${OUT_DIR}"/out_pe_art_perl_R1.txt
 cmp "${OUT_DIR}"/out_pe_art_cxx_sam_R2.txt "${OUT_DIR}"/out_pe_art_perl_R2.txt
 
-rm -fr "${OUT_DIR}"/*
+APB_EXEC \
+    --i-file "${OUT_DIR}"/out_pe.sam \
+    --read_len_1 "${RLEN}" \
+    --read_len_2 "${RLEN}" \
+    --is_pe \
+    --o-file1 "${OUT_DIR}"/out_pe_art_cxx_sam_R1.txt \
+    --o-file2 "${OUT_DIR}"/out_pe_art_cxx_sam_R2.txt \
+    --parallel "${PARALLEL}" \
+    --i-num_threads 4 \
+    --old_behavior
+
+set -x
+[ "$(grep -c -v '^$' <"${OUT_DIR}"/out_pe_art_cxx_sam_R1.txt)" == "432" ]
+[ "$(grep -c -v '^$' <"${OUT_DIR}"/out_pe_art_cxx_sam_R2.txt)" == "432" ]
+set +x
+
+APB_EXEC \
+    --i-file "${OUT_DIR}"/out_pe.sam \
+    --read_len_1 10 \
+    --read_len_2 "${RLEN}" \
+    --is_pe \
+    --o-file1 "${OUT_DIR}"/out_pe_art_cxx_sam_R1.txt \
+    --o-file2 "${OUT_DIR}"/out_pe_art_cxx_sam_R2.txt \
+    --parallel "${PARALLEL}" \
+    --i-num_threads 4 \
+    --old_behavior
+
+set -x
+[ "$(grep -c -v '^$' <"${OUT_DIR}"/out_pe_art_cxx_sam_R1.txt)" == "120" ]
+[ "$(grep -c -v '^$' <"${OUT_DIR}"/out_pe_art_cxx_sam_R2.txt)" == "432" ]
+set +x
+
+rm -fr "${OUT_DIR:?}"/*
 assert_cleandir
+
 unset RLEN
